% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticPricers.R
\name{analyticPricer}
\alias{analyticPricer}
\title{Analytic European option pricing under three basic models}
\usage{
analyticPricer(strike, expiry, spot, model, type = "call")
}
\arguments{
\item{strike}{a single strike price}

\item{expiry}{a single expiry/maturity in YTE}

\item{spot}{the current spot price of the underlying}

\item{model}{the dynamics defining the model, see details}

\item{type}{the type of option to price}
}
\value{
numeric
}
\description{
{Compute European option prices under three basic models: Black-Scholes,
Merton's jump diffusion, and a log-normal mixture.}
}
\details{
{The argument \code{model} must be a named list of
\itemize{
\item \code{name} either "gbm", "gcpp", "mixture", or "merton"
\item \code{param} the parameters defining the above model.}
For "gbm" and "merton", \code{param} should be a vector of the risk-free rate,
volatility, and the same with the mean rate of jumps and jump parameters. For
"mixture" it must be a matrix of probabilities, risk-neutral rate, and volatilities. For
"gcpp", it should be a vector of the risk-free rate, jump rate, jump size, and compensator size.}
}
