% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdePricer.R
\name{bs_greeks}
\alias{bs_greeks}
\title{Compute Greeks under BS dynamics}
\usage{
bs_greeks(
  strike,
  expiry,
  spot,
  v,
  r,
  type = "call",
  N = 100,
  M = 100,
  american = TRUE
)
}
\arguments{
\item{strike}{the strike price}

\item{expiry}{the time until maturity}

\item{spot}{the spot price}

\item{v}{volatility}

\item{r}{risk-free rate}

\item{type}{put or call}

\item{N}{time resolution}

\item{M}{space resolution}

\item{american}{boolean for American or European options}
}
\value{
vector
}
\description{
{The basic Greeks under Black-Scholes.}
}
