% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{callDebitSpread}
\alias{callDebitSpread}
\title{The terminal payoff of a call debit spread}
\usage{
callDebitSpread(spot, strikes)
}
\arguments{
\item{spot}{the current spot price}

\item{strikes}{a vector of two strikes in ascending order}
}
\value{
numeric or vector
}
\description{
The terminal payoff of a call debit spread
}
