% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{callOption}
\alias{callOption}
\title{The terminal payoff of a call option}
\usage{
callOption(spot, strike)
}
\arguments{
\item{spot}{the current spot price}

\item{strike}{the strike price of the option}
}
\value{
numeric or vector
}
\description{
The terminal payoff of a call option
}
