% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impliedParmeters.R
\name{impliedRate}
\alias{impliedRate}
\title{Back out the implied volatility of an option given its market price}
\usage{
impliedRate(strike, maturity, spot, price, volat, type, N, M, ub = 5)
}
\arguments{
\item{strike}{the strike price of the option to backout parameters for}

\item{maturity}{the maturity of the option, in trading years}

\item{spot}{the spot price of the underlying}

\item{price}{the market price of the option contract}

\item{volat}{the risk-free rate}

\item{type}{the type of option: put or call}

\item{N}{number of time sub-intervals}

\item{M}{number of space sub-intervals}

\item{ub}{the bound on the risk-free rate search-space}
}
\value{
numeric
}
\description{
Back out the implied volatility of an option given its market price
}
