% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impliedParmeters.R
\name{ir_error}
\alias{ir_error}
\title{Signed IR error}
\usage{
ir_error(r, strike, maturity, spot, price, volat, type, N, M)
}
\arguments{
\item{r}{the risk-free rate guess}

\item{strike}{the strike price of the option to backout parameters for}

\item{maturity}{the maturity of the option, in trading years}

\item{spot}{the spot price of the underlying}

\item{price}{the market price of the option contract}

\item{volat}{the volatility level (annualized)}

\item{type}{the type of option: put or call}

\item{N}{number of time sub-intervals}

\item{M}{number of space sub-intervals}
}
\value{
numeric
}
\description{
Signed IR error
}
