% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impliedParmeters.R
\name{iv_error}
\alias{iv_error}
\title{Signed IV error}
\usage{
iv_error(v, strike, maturity, spot, price, rate, type, N, M)
}
\arguments{
\item{v}{the volatility guess}

\item{strike}{the strike price of the option to backout parameters for}

\item{maturity}{the maturity of the option, in trading years}

\item{spot}{the spot price of the underlying}

\item{price}{the market price of the option contract}

\item{rate}{the risk-free rate}

\item{type}{the type of option: put or call}

\item{N}{number of time sub-intervals}

\item{M}{number of space sub-intervals}
}
\value{
numeric
}
\description{
Signed IV error
}
