% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarloPricer.R
\name{monteCarlo}
\alias{monteCarlo}
\title{A generic Monte-Carlo integrator}
\usage{
monteCarlo(maturity, rate, variates)
}
\arguments{
\item{maturity}{the maturity of the option}

\item{rate}{the discounting rate}

\item{variates}{the simulated terminal payoff variates to average}
}
\value{
numeric
}
\description{
{A generic Monte-Carlo integrator given a vector of
simulated terminal payoff variates.}
}
