% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarloPricer.R
\name{pricer_mc}
\alias{pricer_mc}
\title{Monte-Carlo European option pricing}
\usage{
pricer_mc(strikes, expiries, spot, model, type = "call", n = 1000)
}
\arguments{
\item{strikes}{vector of strike prices}

\item{expiries}{vector of maturities, in trading years}

\item{spot}{the current spot price of the underlying}

\item{model}{the dynamics defining the model, see details}

\item{type}{the type of option to price}

\item{n}{number of variates in Monte-Carlo integrator}
}
\value{
data.frame
}
\description{
{Compute European option prices under three basic models: Black-Scholes,
Merton's jump diffusion, and a log-normal mixture.}
}
\details{
{The argument \code{model} must be a named list of
\itemize{
\item \code{name} either "gbm", "mixture", or "merton"
\item \code{param} the parameters defining the above model.}
For "gbm" and "merton", \code{param} should be a vector of the risk-free rate,
volatility, and the same with the mean rate of jumps and jump parameters. For
"mixture" it must be a matrix of probabilities, risk-neutral rate, and volatilities.}
}
