% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdePricer.R
\name{pricer_pde}
\alias{pricer_pde}
\title{Option pricing via PDE finite-difference solvers}
\usage{
pricer_pde(
  strikes,
  expiries,
  spot,
  model,
  type = "call",
  N = 100,
  M = 100,
  american = TRUE
)
}
\arguments{
\item{strikes}{vector of strike prices}

\item{expiries}{vector of maturities, in trading years}

\item{spot}{the current spot price of the underlying}

\item{model}{the dynamics defining the model, see details}

\item{type}{the type of option to price}

\item{N}{time-resolution}

\item{M}{space-resolution}

\item{american}{boolean for American options (TRUE) or European options}
}
\value{
data.frame
}
\description{
{Compute European and American option prices under three basic models: Black-Scholes, and a log-normal mixture. Use
the PIDE solver to price under Merton's jump-diffusion.}
}
\details{
{The argument \code{model} must be a named list of
\itemize{
\item \code{name} either "gbm" or "mixture"
\item \code{param} the parameters defining the above model.}
For "gbm", \code{param} should be a vector of the risk-free rate,
volatility, and the same with the mean rate of jumps and jump parameters. For
"mixture" it must be a matrix of probabilities, risk-neutral rate, and volatilities.}
}
