% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing-package.R
\docType{package}
\name{pricing-package}
\alias{pricing}
\alias{pricing-package}
\title{pricing: Option pricing engines for (jump)-diffusions}
\description{
European and American option pricing, hedging, and calibration.
    Models provided are Black-Scholes, Brigo-Mercurio Mixture, and Merton's
    jump diffusion. Monte-Carlo and finite-difference solvers are available
    for European options while just the latter is provided to price American 
    options.
}
\keyword{internal}
