% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{putDebitSpread}
\alias{putDebitSpread}
\title{The terminal payoff of a put debit spread}
\usage{
putDebitSpread(spot, strikes)
}
\arguments{
\item{spot}{the current spot price}

\item{strikes}{a vector of two strikes in ascending order}
}
\value{
numeric or vector
}
\description{
The terminal payoff of a put debit spread
}
