% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{putOption}
\alias{putOption}
\title{The terminal payoff of a put option}
\usage{
putOption(spot, strike)
}
\arguments{
\item{spot}{the current spot price}

\item{strike}{the strike price of the option}
}
\value{
numeric or vector
}
\description{
The terminal payoff of a put option
}
