% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tridiagonalSolvers.R
\name{trisolveR}
\alias{trisolveR}
\title{Thomas-algorithm for tridiagonal linear systems}
\usage{
trisolveR(alpha, beta, delta, b.vec)
}
\arguments{
\item{alpha}{lower diagonal}

\item{beta}{diagonal}

\item{delta}{upper diagonal}

\item{b.vec}{target vector}
}
\value{
numeric vector
}
\description{
{Solve tridiagonal linear systems with the Thomas-algorithm.}
}
